feat(bronze): implement exception handling and refactor script names and SQL changes
[ADD]
- Added exception handling in python scripts to improve error management during execution.

[UPDATE]
- Renamed python scripts for better clarity and maintainability.
- Updated ddl_bronze.sql to change 'prd_cost' column type from int to float to accommodate more precise values.
- Renamed producer_info.csv to product_info.csv.

+++ scripts/sql/bronze/ddl_bronze.sql
@@ -27,7 +27,7 @@ CREATE TABLE bronze.crm_prd_info (
     prd_id       INT,
     prd_key      VARCHAR(50),
     prd_nm       VARCHAR(50),
-    prd_cost     INT,
+    prd_cost     FLOAT,
     prd_line     VARCHAR(50),
     prd_start_dt DATE,
     prd_end_dt   DATE
